requirements.txt
ccxt
streamlit
pandas
plotly
pandas_ta
python-dotenv
fastapi
uvicorn
psycopg2-binary
